+++ tests/des/vectors/vectors_des.h
@@ -1,0 +1,23 @@
+/*
+ * Copyright 2023-2024 The Gestalt Project Authors. All Rights Reserved.
+ *
+ * Licensed under the MIT License. See the file LICENSE for the full text.
+ */
+
+/*
+ * vectors_des.cpp
+ *
+ * This file contains the vectors used to test DES.
+ */
+
+#pragma once
+
+const std::string key = "752878397493CB70";
+const std::string key2 = "10316E028C8F3B4A";
+const std::string key3 = "7CA110454A1A6E57";
+const std::string nonce = "0102030405060708";
+const std::string plaintext = "Hello, Gestalt!";
+const std::string multiBlockPT = 
+    "In the spring of her twenty-second year, Sumire fell in love for the first time in her life. An intense love, "
+    "a veritable tornado sweeping across the plains flattening everything in its path, tossing things up in the air, "
+    "ripping them to shreds, crushing them to bits.";
