+++ tests/ecc/vectors/vectors_ecdh.h
@@ -5,18 +5,9 @@
  */
 
 /*
- * ecdhTests.cpp
+ * vectors_ecdh.cpp
  *
- * This file containts the unit tests for the ECDH (Elliptic Curve Diffie Hellman Algorithm) Gestalt implementation. 
- * 
  */
-
-#include <gestalt/ecdh.h>
-
-#include "gtest/gtest.h"
-#include <vector>
-#include <string>
-#include <iostream>
 
 /* Test vectors from: https://csrc.nist.gov/Projects/cryptographic-algorithm-validation-program/Component-Testing */
 static const struct ECDHTestVector {
@@ -91,78 +82,3 @@ class ECDHComputeSharedSecret : public testing::TestWithParam<ECDHTestVector> {
 };
 
 INSTANTIATE_TEST_SUITE_P(All, ECDHComputeSharedSecret, testing::ValuesIn(kECDHTestVectors), CustomNameGenerator);
-
-TEST_P(ECDHComputeSharedSecret, computeSharedSecret)
-{
-    const ECDHTestVector &test = GetParam();
-    SCOPED_TRACE(test.name);
-
-    ECDH ecdh(test.curve, test.dIUT);
-    
-    Point peerPublicKey(test.QCAVSx, test.QCAVSy);
-    ecdh.getPublicKey(peerPublicKey);
-
-    std::string computedZ = ecdh.computeSharedSecret();
-    std::string expected = test.ZIUT;
-
-    EXPECT_TRUE(computedZ == expected);
-}
-
-TEST(ECDHComputeSharedSecret, computeSharedSecretAsPair)
-{
-    ECDH alice(StandardCurve::P256, "0x7d7dc5f71eb29ddaf80d6214632eeae03d9058af1fb6d22ed80badb62bc1a534");
-    ECDH bob  (StandardCurve::P256, "0x38f65d6dce47676044d58ce5139582d568f64bb16098d179dbab07741dd5caf5");
-
-    std::string aliceSharedSecret = alice.computeSharedSecret(bob.givePublicKey());
-    std::string bobSharedSecret = bob.computeSharedSecret(alice.givePublicKey());
-
-    EXPECT_TRUE(aliceSharedSecret == bobSharedSecret);
-}
-
-class TestECDH : public ::testing::Test {
-private:
-    ECDH ecdh;
-protected:
-    void getPublicKey(const Point& givenPublicKey) { ecdh.peerPublicKey = givenPublicKey; };
-    Point retrievePublicKey() { return ecdh.peerPublicKey; };
-    std::string keyToString(const Point& point) const { return ecdh.keyToString(point); };
-};
-
-TEST_F(TestECDH, keyToString)
-{
-    Point P("0xCEC028EE08D09E02672A68310814354F9EABFFF0DE6DACC1CD3A774496076AE", 
-            "0xEFF471FBA0409897B6A48E8801AD12F95D0009B753CF8F51C128BF6B0BD27FBD");
-    
-    std::string result = keyToString(P);
-
-    std::string expected = "cec028ee08d09e02672a68310814354f9eabfff0de6dacc1cd3a774496076ae";
-
-    EXPECT_EQ(result, expected);
-}
-
-TEST_F(TestECDH, givePublicKey)
-{
-    ECDH alice("0x519B423D715F8B581F4FA8EE59F4771A5B44C8130B4E3EACCA54A56DDA72B464");
-
-    Point givenPubKey = alice.givePublicKey();
-
-    Point expected("0xCEC028EE08D09E02672A68310814354F9EABFFF0DE6DACC1CD3A774496076AE", 
-                   "0xEFF471FBA0409897B6A48E8801AD12F95D0009B753CF8F51C128BF6B0BD27FBD");
-
-    EXPECT_TRUE(mpz_cmp(givenPubKey.x, expected.x) == 0);
-    EXPECT_TRUE(mpz_cmp(givenPubKey.y, expected.y) == 0);
-}
-
-TEST_F(TestECDH, getPublicKey)
-{
-    ECDH bob("0x519B423D715F8B581F4FA8EE59F4771A5B44C8130B4E3EACCA54A56DDA72B464");
-
-    getPublicKey(bob.givePublicKey());
-    Point givenPubKey = retrievePublicKey();
-
-    Point expected("0xCEC028EE08D09E02672A68310814354F9EABFFF0DE6DACC1CD3A774496076AE", 
-                   "0xEFF471FBA0409897B6A48E8801AD12F95D0009B753CF8F51C128BF6B0BD27FBD");
-
-    EXPECT_TRUE(mpz_cmp(givenPubKey.x, expected.x) == 0);
-    EXPECT_TRUE(mpz_cmp(givenPubKey.y, expected.y) == 0);
-}
